Keep profiles in storage.profiles
Before this update, the localStorage contained only the profile
of the current user.
Now it contains the profiles of all the users who had used the
app.

It will be used when we'll keep data about the courses and
exercices. It will make users able to come back to the app
and restart from the point where the leaved last time.